requirements.txt
asgiref==3.8.1
attrs==24.2.0
cffi==1.17.1
crispy-bootstrap5==2024.10
cryptography==44.0.0
Django==5.1.3
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-dbbackup==4.2.1
django-encrypted-model-fields==0.6.5
django-filter==24.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-generators==0.5.0
drf-spectacular==0.27.2
et_xmlfile==2.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
openpyxl==3.1.5
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
